BF: relax minimum part size for tests
multipart upload related unit tests rely on the API functions to catch
error in a certain order. This is no longer the case for some reason
which I can't understand.

Unit tests must be fixed and this temporary fix must be reverted.
When building a fake request to test a specific error. Make sure
that the request is only affected by this specific error.